requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
ib_insync==0.9.86
SQLAlchemy==2.0.30
scipy==1.13.0
python-dotenv==1.0.1
